// src/types/mentorHashtag.ts
export type HASHTAG =
  | '전체'
  | '#유학'
  | '#인공지능'
  | '#연세대'
  | '#재료'
  | '#삼성전자'
  | '#데이터사이언스'
  | '#영문교정/컨설팅';

export const HASHTAGS_LIST: HASHTAG[] = [
  '전체',
  '#유학',
  '#인공지능',
  '#연세대',
  '#재료',
  '#삼성전자',
  '#데이터사이언스',
  '#영문교정/컨설팅'
];